# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Supply Chain – Micro-Warehouse Network for Resilient Logistics (India)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

  # -----------------------------------------------------------------------
  # Locations (warehouses + micro-warehouses)
  # -----------------------------------------------------------------------
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: customerId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place order (auto-routes to nearest source)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "409":
          description: No stock available
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }

  # -----------------------------------------------------------------------
  # Alerts
  # -----------------------------------------------------------------------
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    LocationType:
      type: string
      enum: [warehouse, micro]
...
    OrderStatus:
      type: string
      enum: [pending, dispatched, in_transit, delivered, cancelled]
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        unitPrice: { type: number }
        imageUrl:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
      required: [id, name, sku, category, unitPrice, createdAt]
    CreateProductBody:
      type: object
      properties:
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        unitPrice: { type: number }
        imageUrl:
          type: ["string", "null"]
      required: [name, sku, category, unitPrice]
    UpdateProductBody:
      type: object
      properties:
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        unitPrice: { type: number }
        imageUrl:
          type: ["string", "null"]
...
    Order:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: string }
        productId: { type: integer }
        productName: { type: string }
        productSku: { type: string }
        quantity: { type: integer }
        sourceLocationId:
          type: ["integer", "null"]
        sourceLocationName:
          type: ["string", "null"]
        sourceLocationType:
          oneOf:
            - $ref: "#/components/schemas/LocationType"
            - type: "null"
        deliveryAddress: { type: string }
        deliveryCity: { type: string }
        deliveryLat: { type: number }
        deliveryLng: { type: number }
        distanceKm: { type: number }
        etaMinutes: { type: integer }
        status: { $ref: "#/components/schemas/OrderStatus" }
        totalPrice: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [id, customerId, customerName, productId, productName, productSku,
         quantity, deliveryAddress, deliveryCity, deliveryLat, deliveryLng,
         distanceKm, etaMinutes, status, totalPrice, createdAt, updatedAt]
    CreateOrderBody:
      type: object
      properties:
        customerId: { type: integer }
        productId: { type: integer }
        quantity: { type: integer }
        deliveryAddress: { type: string }
        deliveryCity: { type: string }
        deliveryLat: { type: number }
        deliveryLng: { type: number }
      required:
        [customerId, productId, quantity, deliveryAddress, deliveryCity,
         deliveryLat, deliveryLng]
    UpdateOrderStatusBody:
      type: object
      properties:
        status: { $ref: "#/components/schemas/OrderStatus" }
      required: [status]
...